+++ packages/generator-teams/src/app/manifestGeneration/manifestGenerators/generator18/BotManifestUpdater.ts
@@ -51,11 +51,6 @@ export class BotManifestUpdater implements IManifestUpdater {
             newbot.supportsFiles = true;
         }
 
-        if (options.botCallingEnabled) {
-            newbot.supportsCalling = true;
-            newbot.supportsVideo = true;
-        }
-
         (<any[]>manifest.bots).push(newbot);
     }
 }