+++ packages/generator-teams/src/app/manifestGeneration/manifestGenerators/generatorDevPreview/ManifestGenerator.ts
@@ -32,18 +32,18 @@ export class ManifestGenerator extends BaseManifestGenerator {
     }
 
     public supportsUpdateManifest(from: string): boolean {
-        return from === "1.8" ||
-            from === "1.9" ||
-            from === "1.10" ||
-            from === "1.11";
+        return from === "1.11" ||
+            from === "m365devPreview" ||
+            from === "1.12" ||
+            from === "1.13";
     }
 
     public updateManifest(manifest: any, log?: (message?: string, context?: any) => void): any {
         switch (manifest.manifestVersion) {
-            case "1.8":
-            case "1.9":
-            case "1.10":
             case "1.11":
+            case "1.12":
+            case "1.13":
+            case "m365DevPreview":
                 manifest["$schema"] = "https://raw.githubusercontent.com/OfficeDev/microsoft-teams-app-schema/preview/DevPreview/MicrosoftTeams.schema.json";
                 manifest.manifestVersion = "devPreview";
                 return manifest;