+++ packages/generator-teams/src/tab/TabGenerator.ts
@@ -7,10 +7,8 @@ import * as lodash from 'lodash';
 import { GeneratorTeamsAppOptions } from './../app/GeneratorTeamsAppOptions';
 import { Yotilities } from './../app/Yotilities';
 import { ManifestGeneratorFactory } from '../app/manifestGeneration/ManifestGeneratorFactory';
-import validate = require('uuid-validate');
-import EmptyGuid = require('../app/EmptyGuid');
-import { ManifestVersions } from '../app/manifestGeneration/ManifestVersions';
-import { parse } from "tldjs";
+import * as validate from 'uuid-validate';
+import * as EmptyGuid from '../app/EmptyGuid';
 
 export const OFFICE_GUID = "00000003-0000-0ff1-ce00-000000000000";
 
@@ -65,29 +63,11 @@ export class TabGenerator extends Generator {
                                 }
                             },
                             {
-                                name: "Personal (static)",
+                                name: "Personal (for Teams, Outlook and Office)",
                                 value: "static",
                                 disabled: () => {
                                     // only 16 static tabs are allowed
                                     return this.options.existingManifest && this.options.existingManifest.configurableTabs && this.options.existingManifest.configurableTabs.length > 16;
-                                }
-                            },
-                            {
-                                name: "Viva Connections (static tab)",
-                                value: "viva",
-                                disabled: (answers) => {
-                                    if (this.options.manifestVersion == ManifestVersions.v18) {
-                                        return true; // don't enable for v1.8
-                                    }
-                                    if ((this.options.existingManifest &&
-                                        this.options.existingManifest.webApplicationInfo)) {
-                                        // don't enable Viva if there's SSO configured already
-                                        return true;
-                                    }
-                                    // only 16 static tabs are allowed
-                                    return this.options.existingManifest &&
-                                        this.options.existingManifest.configurableTabs &&
-                                        this.options.existingManifest.configurableTabs.length > 16;
                                 }
                             }
                         ],
@@ -126,9 +106,6 @@ export class TabGenerator extends Generator {
                         prefix: generatorPrefix,
                         default: false,
                         when: (answers: any) => {
-                            if (answers.tabType == "viva") {
-                                return false; // SSO will automatically be configured for Viva
-                            }
                             return (this.options.existingManifest === undefined ||
                                 this.options.existingManifest &&
                                 this.options.existingManifest.webApplicationInfo == undefined);
@@ -190,25 +167,7 @@ export class TabGenerator extends Generator {
                         when: (answers: any) => {
                             return answers.tabSharePoint;
                         }
-                    },
-                    {
-                        type: 'input',
-                        name: 'vivaUrl',
-                        message: 'Enter the link of the SharePoint portal that you want to pin in Teams.',
-                        prefix: generatorPrefix,
-                        validate: (input: string) => {
-                            const domain = parse(input);
-                            if (!domain.isValid) {
-                                return "Please enter a valid domain"
-                            }
-                            if (domain.domain !== "sharepoint.com") {
-                                return "It has to be a sharepoint.com domain";
-                            }
-                            return true;
-                        },
-                        when: (answers: any) => answers.tabType == "viva",
-                    },
-
+                    }
                 ]
             ).then((answers: any) => {
                 this.options.tabTitle = answers.tabTitle;
@@ -226,64 +185,57 @@ export class TabGenerator extends Generator {
                 this.options.tabSSO = answers.tabSSO;
                 this.options.tabSSOAppId = answers.tabSSOAppId;
                 this.options.tabSSOAppUri = answers.tabSSOAppUri;
-                this.options.vivaUrl = answers.vivaUrl;
-                if (this.options.vivaUrl && this.options.vivaUrl.endsWith("/")) {
-                    this.options.vivaUrl = this.options.vivaUrl.substr(0, this.options.vivaUrl.length - 1);
-                }
-                if(this.options.tabType == "viva") {
-                    this.options.isFullScreen = true;
-                }
             });
         }
     }
     public writing() {
         if (this.options.tab) {
-            if (this.options.tabType !== "viva") {
-                let templateFiles = [
-                    "src/client/{tabName}/{tabReactComponentName}.tsx",
-                    "src/server/{tabName}/{tabReactComponentName}.ts",
-                    "src/public/{tabName}/index.html",
-                ];
+
+            let templateFiles = [
+                "src/client/{tabName}/{tabReactComponentName}.tsx",
+                "src/server/{tabName}/{tabReactComponentName}.ts",
+                "src/public/{tabName}/index.html",
+            ];  
+            if (this.options.tabType == "configurable") {
+                templateFiles.push(
+                    "src/client/{tabName}/{tabReactComponentName}Config.tsx",
+                    "src/client/{tabName}/{tabReactComponentName}Remove.tsx",
+                    "src/public/{tabName}/remove.html",
+                    "src/public/{tabName}/config.html",
+                );
+            }
+
+            if (this.options.unitTestsEnabled) {
+                templateFiles.push(
+                    "src/client/{tabName}/__tests__/{tabReactComponentName}.spec.tsx",
+                );
                 if (this.options.tabType == "configurable") {
                     templateFiles.push(
-                        "src/client/{tabName}/{tabReactComponentName}Config.tsx",
-                        "src/client/{tabName}/{tabReactComponentName}Remove.tsx",
-                        "src/public/{tabName}/remove.html",
-                        "src/public/{tabName}/config.html",
+                        "src/client/{tabName}/__tests__/{tabReactComponentName}Config.spec.tsx",
+                        "src/client/{tabName}/__tests__/{tabReactComponentName}Remove.spec.tsx",
                     );
-                }
-
-                if (this.options.unitTestsEnabled) {
-                    templateFiles.push(
-                        "src/client/{tabName}/__tests__/{tabReactComponentName}.spec.tsx",
-                    );
-                    if (this.options.tabType == "configurable") {
-                        templateFiles.push(
-                            "src/client/{tabName}/__tests__/{tabReactComponentName}Config.spec.tsx",
-                            "src/client/{tabName}/__tests__/{tabReactComponentName}Remove.spec.tsx",
-                        );
-                    }
-                }
-
-                this.sourceRoot()
-
-                templateFiles.forEach(t => {
-                    this.fs.copyTpl(
-                        this.templatePath(t),
-                        Yotilities.fixFileNames(t, this.options),
-                        this.options);
-                });
-
-                if (this.options.tabSharePoint) {
-                    const from = "src/public/assets/tab-preview.png";
-                    const to = "src/public/assets/{tabName}-preview.png";
-                    this.fs.copy(
-                        this.templatePath(from),
-                        Yotilities.fixFileNames(to, this.options),
-                        {},
-                        this.options);
                 }
             }
+
+            this.sourceRoot();
+
+            templateFiles.forEach(t => {
+                this.fs.copyTpl(
+                    this.templatePath(t),
+                    Yotilities.fixFileNames(t, this.options),
+                    this.options);
+            });
+
+            if (this.options.tabSharePoint) {
+                const from = "src/public/assets/tab-preview.png";
+                const to = "src/public/assets/{tabName}-preview.png";
+                this.fs.copy(
+                    this.templatePath(from),
+                    Yotilities.fixFileNames(to, this.options),
+                    {},
+                    this.options);
+            }
+
 
             // Update manifest
             const manifestGeneratorFactory = new ManifestGeneratorFactory();
@@ -295,46 +247,45 @@ export class TabGenerator extends Generator {
 
             this.fs.writeJSON(manifestPath, manifest);
 
-            if (this.options.tabType !== "viva") {
-                if (this.options.tabSSO) {
-                    Yotilities.addAdditionalDeps([
-                        ["jwt-decode", "^3.1.2"]
-                    ], this.fs);
+            if (this.options.tabSSO) {
+                Yotilities.addAdditionalDeps([
+                    ["jwt-decode", "^3.1.2"]
+                ], this.fs);
 
-                    // update .env file
-                    Yotilities.addOrUpdateEnv(".env", `TAB_APP_ID`, this.options.tabSSOAppId, this.fs);
-                    Yotilities.addOrUpdateEnv(".env", `TAB_APP_URI`, this.options.tabSSOAppUri, this.fs);
-                }
+                // update .env file
+                Yotilities.addOrUpdateEnv(".env", `TAB_APP_ID`, this.options.tabSSOAppId, this.fs);
+                Yotilities.addOrUpdateEnv(".env", `TAB_APP_URI`, this.options.tabSSOAppUri, this.fs);
+            }
 
-                // update client.ts
+            // update client.ts
+            Yotilities.insertTsExportDeclaration(
+                "src/client/client.ts",
+                `./${this.options.tabName}/${this.options.tabReactComponentName}`,
+                `Automatically added for the ${this.options.tabName} tab`,
+                this.fs
+            );
+            if (this.options.tabType == "configurable") {
                 Yotilities.insertTsExportDeclaration(
                     "src/client/client.ts",
-                    `./${this.options.tabName}/${this.options.tabReactComponentName}`,
-                    `Automatically added for the ${this.options.tabName} tab`,
+                    `./${this.options.tabName}/${this.options.tabReactComponentName}Config`,
+                    undefined,
                     this.fs
                 );
-                if (this.options.tabType == "configurable") {
-                    Yotilities.insertTsExportDeclaration(
-                        "src/client/client.ts",
-                        `./${this.options.tabName}/${this.options.tabReactComponentName}Config`,
-                        undefined,
-                        this.fs
-                    );
-                    Yotilities.insertTsExportDeclaration(
-                        "src/client/client.ts",
-                        `./${this.options.tabName}/${this.options.tabReactComponentName}Remove`,
-                        undefined,
-                        this.fs
-                    );
-                }
-
                 Yotilities.insertTsExportDeclaration(
-                    "src/server/TeamsAppsComponents.ts",
-                    `./${this.options.tabName}/${this.options.tabReactComponentName}`,
-                    `Automatically added for the ${this.options.tabName} tab`,
+                    "src/client/client.ts",
+                    `./${this.options.tabName}/${this.options.tabReactComponentName}Remove`,
+                    undefined,
                     this.fs
                 );
             }
+
+            Yotilities.insertTsExportDeclaration(
+                "src/server/TeamsAppsComponents.ts",
+                `./${this.options.tabName}/${this.options.tabReactComponentName}`,
+                `Automatically added for the ${this.options.tabName} tab`,
+                this.fs
+            );
+
         }
     }
 }