+++ packages/generator-teams/tests/BotGenerator.spec.ts
@@ -54,19 +54,6 @@ describe("teams:bot", async () => {
         });
       });
     }
-    if (prompts.botCallingEnabled) {
-      it("Manifest has calling support", async () => {
-        assert.jsonFileContent("src/manifest/manifest.json", {
-          bots: [{ supportsCalling: true, supportsVideo: true }]
-        });
-      });
-    } else {
-      it("Manifest should not have calling support", async () => {
-        assert.noJsonFileContent("src/manifest/manifest.json", {
-          bots: [{ supportsCalling: true, supportsVideo: true }]
-        });
-      });
-    }
 
     if (prompts.unitTestsEnabled && prompts.staticTab) {
       it("Should have bot unit test files", async () => {
@@ -84,7 +71,7 @@ describe("teams:bot", async () => {
       });
       it("Should have package botbuilder-testing", async () => {
         assert.jsonFileContent("package.json", {
-          devDependencies: { "botbuilder-testing": "4.14.1" }
+          devDependencies: { "botbuilder-testing": "~4.16.0" }
         });
       });
     } else {
@@ -93,7 +80,7 @@ describe("teams:bot", async () => {
       });
       it("Should not have package botbuilder-testing", async () => {
         assert.noJsonFileContent("package.json", {
-          devDependencies: { "botbuilder-testing": "4.14.1" }
+          devDependencies: { "botbuilder-testing": "~4.16.0" }
         });
       });
     }