+++ packages/yoteams-deploy/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [*1.2.0*]- <*2022-06-23*>
+
+### Changes
+
+* Requires Node 14 or later
+* Dependency package updates
+
 ## [*1.1.0*]- <*2021-11-18*>
 
 ### Added